Move test files to new folders
